fix: add backup check and create backups directory in update_kubeconfg.sh

+++ dotenv/kube/update_kubeconfg.sh
@@ -9,8 +9,11 @@ kube_dir=~/.kube
 cd $kube_dir/config.d
 kubeconfig_str=$(ls -1rt *.config | xargs | sed s/\ /:/g)
 
-echo "INFO: Backing up the config to directory ${kube_dir}/backups..."
-cp -p ${kube_dir}/config ${kube_dir}/backups/config.$(date +%Y%m%d%H%M)
+if [[ -f ${kube_dir}/config ]]; then
+  echo "INFO: Backing up the config to directory ${kube_dir}/backups..."
+  mkdir -p ${kube_dir}/backups
+  cp -p ${kube_dir}/config ${kube_dir}/backups/config.$(date +%Y%m%d%H%M)
+fi
 
 echo "INFO: Regenerating the ${kube_dir}/config..."
 KUBECONFIG=${kubeconfig_str} ${kube_cmd} config view --merge --flatten > ${kube_dir}/config
